Correctly implement helper function and unit tests
